requirements.txt
fastapi[standart]
fastapi-sqlalchemy
sqlalchemy[asyncio]
asyncpg
python-multipart
jinja2
passlib
bcrypt
pyjwt

